Increased change of direction speed and switching modes
Updated Pin 34, it is now checking to see if the Pin's value is 0. A Manual switch on the controller connects Pin 34 to ground when switched into on position.

Reduced the range for entering change of direction to >50 and <3150.

Added stop to all directional movements to all none needed motors, this allows for changes of direction to happen instantly. Previously if you changed directly from Forwards to Left, it would have conflicting motor's in "on()" position so they would stall. Now that's not possible as it turns all motors not need to move in that direction off before it moves.